Rename: images.model.ts -> image.model.ts, tags.model.ts -> tag.model.ts 파일명 변경
