+++ server/model/image.model.ts
